Unsplit PSP to match
