+++ src/dra_psp/507F0.c
@@ -2194,3 +2194,38 @@ void func_80124164(
         prim->b0 = prim->b2 = D_800AE290 * colorIntensity / 0x100;
     }
 }
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/507F0", EntityTeleport);
+
+// Entity #66. Made by final blueprint, #122. Created in PlayerStepStand.
+void EntityPlayerSleepZ(Entity* self) {
+    if (PLAYER.step != Player_Stand || PLAYER.step_s != Player_Stand_ChairSit) {
+        DestroyEntity(self);
+        return;
+    }
+
+    switch (self->step) {
+    case 0:
+        self->animSet = ANIMSET_DRA(17);
+        self->velocityY = FIX(-0.375);
+        SetSpeedX(FIX(0.25));
+        self->unk5A = 0x50;
+        self->palette = PAL_OVL(0x19F);
+        self->anim = D_800AE294;
+
+        self->posY.i.hi -= 16;
+        self->posX.val += self->velocityX << 5;
+        self->flags = FLAG_UNK_100000;
+        self->facingLeft = 0;
+        self->step++;
+        break;
+
+    case 1:
+        self->posX.val += self->velocityX;
+        self->posY.val += self->velocityY;
+        if (self->animFrameDuration < 0) {
+            DestroyEntity(self);
+        }
+        break;
+    }
+}
