+++ src/dra_psp/59E20.c
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../dra/dra.h"
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/59E20", func_psp_091367A0);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/59E20", func_psp_09136BF8);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/59E20", func_psp_09136D18);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/59E20", func_psp_09136E18);
